Added Date Toggle
Done: Date toggles, sets or unsets date when changed.
Datepicker sets date

ToDo:
There might still be double state with hasDueDate, but I don't know how to get around that with Toggle
Use tapgesture to minimize or show datepicker
when datepicker is minimized, show text next to or under date of what the due date is
extend application to time
only allow time to be selected if date is already selected, or make it toggle on.
when date is toggled off, also toggle off time.