Date toggle is working
Done: Due Date toggle is working.
Calendar expands when toggle is turned on, collapses when turned off.
User can tap on date to open calendar if turned on.

Next: Figure out if time and date work separately or together
Create another toggle for time that auto-turns on date if selected